+++ docs/source/release-notes.rst
@@ -2,6 +2,43 @@
 Release Notes
 ==================================
 *************
+Version 0.6.1
+*************
+Release Date xx/xx/xxxx
+
+- Fix initial serial number check to properly raise exception on incorrect entry.
+- Failure to open a drivers pyvisa resource will now raise a more informative InstrumentOpenError exception from the pyvisa error
+- New DMM driver added for Keithley 6500 DMM. 
+- Remove unused error_cleanup functinon and flags in fluke DMM driver.
+- Remove temperature, ftemperature, analog filter and digital filter from DMM drivers.
+- Add _voltage_range to frequency and period measurement modes.
+- Add optional formatting argument ('fmt') to check functions to improve ui display and logging of test values
+- Fix bug where sequencer would return success if terminated during startup
+
+
+*************
+Version 0.6.0
+*************
+Release Date 21/10/2022
+
+Major Changes
+################
+Relocation of configuration files:
+
+- Logs now located in platformdirs.user_log_dir("Fixate")
+- Instrument config file renamed from local_config.json -> instruments.json and located in platformdirs.site_config_dir("Fixate")
+- fixate.yml also moved to platformdirs.site_config_dir("Fixate")
+
+New Features
+############
+- Added gif support for the gui. Try user_gif("<path to .gif>")
+- Clearing the gui image scene will now create a new instance to reset auto-scaling.
+
+Improvements
+############
+- Updated siglent PPS driver to support parsing of error strings in the latest firmware
+
+*************
 Version 0.5.7
 *************
 Release Date 12/07/2022